requirements.txt
Django==6.0.2
djangorestframework==3.16.1
psycopg2-binary==2.9.11
redis==6.4.0
celery[redis]==5.6.2
django-redis==6.0.0
faker==40.1.2
drf-spectacular==0.27.1